consolidate patterns
